+++ medusa/server/api/v2/config.py
@@ -45,6 +45,7 @@ from medusa.server.api.v2.base import (
     iter_nested_items,
     set_nested_value,
 )
+from medusa.show.recommendations.trakt import TraktPopular
 
 from six import iteritems, itervalues, text_type
 from six.moves import map
@@ -108,6 +109,7 @@ class ConfigHandler(BaseRequestHandler):
 
         'indexerDefaultLanguage': StringField(app, 'INDEXER_DEFAULT_LANGUAGE'),
         'showUpdateHour': IntegerField(app, 'SHOWUPDATE_HOUR'),
+        'recommendedShowUpdateHour': IntegerField(app, 'RECOMMENDED_SHOW_UPDATE_HOUR'),
         'indexerTimeout': IntegerField(app, 'INDEXER_TIMEOUT'),
         'indexerDefault': IntegerField(app, 'INDEXER_DEFAULT'),
         'plexFallBack.enable': BooleanField(app, 'FALLBACK_PLEX_ENABLE'),
@@ -183,6 +185,13 @@ class ConfigHandler(BaseRequestHandler):
         'recentShows': ListField(app, 'SHOWS_RECENT'),
         'providers.prowlarr.url': StringField(app, 'PROWLARR_URL'),
         'providers.prowlarr.apikey': StringField(app, 'PROWLARR_APIKEY'),
+
+        'recommended.cache.shows': BooleanField(app, 'CACHE_RECOMMENDED_SHOWS'),
+        'recommended.cache.trakt': BooleanField(app, 'CACHE_RECOMMENDED_TRAKT'),
+        'recommended.cache.imdb': BooleanField(app, 'CACHE_RECOMMENDED_IMDB'),
+        'recommended.cache.anidb': BooleanField(app, 'CACHE_RECOMMENDED_ANIDB'),
+        'recommended.cache.anilist': BooleanField(app, 'CACHE_RECOMMENDED_ANILIST'),
+        'recommended.trakt.selectedLists': ListField(app, 'CACHE_RECOMMENDED_TRAKT_LISTS'),
 
         # Sections
         'clients.torrents.authType': StringField(app, 'TORRENT_AUTH_TYPE'),
@@ -652,6 +661,9 @@ class DataGenerator(object):
 
         section_data['indexerDefaultLanguage'] = app.INDEXER_DEFAULT_LANGUAGE
         section_data['showUpdateHour'] = int_default(app.SHOWUPDATE_HOUR, app.DEFAULT_SHOWUPDATE_HOUR)
+        section_data['recommendedShowUpdateHour'] = int_default(
+            app.RECOMMENDED_SHOW_UPDATE_HOUR, app.DEFAULT_RECOMMENDED_SHOW_UPDATE_HOUR
+        )
         section_data['indexerTimeout'] = int_default(app.INDEXER_TIMEOUT, 20)
         section_data['indexerDefault'] = app.INDEXER_DEFAULT
 
@@ -659,6 +671,15 @@ class DataGenerator(object):
         section_data['plexFallBack']['enable'] = bool(app.FALLBACK_PLEX_ENABLE)
         section_data['plexFallBack']['notifications'] = bool(app.FALLBACK_PLEX_NOTIFICATIONS)
         section_data['plexFallBack']['timeout'] = int(app.FALLBACK_PLEX_TIMEOUT)
+
+        section_data['recommended'] = {'cache': {}, 'trakt': {}}
+        section_data['recommended']['cache']['shows'] = bool(app.CACHE_RECOMMENDED_SHOWS)
+        section_data['recommended']['cache']['trakt'] = bool(app.CACHE_RECOMMENDED_TRAKT)
+        section_data['recommended']['cache']['imdb'] = bool(app.CACHE_RECOMMENDED_IMDB)
+        section_data['recommended']['cache']['anidb'] = bool(app.CACHE_RECOMMENDED_ANIDB)
+        section_data['recommended']['cache']['anilist'] = bool(app.CACHE_RECOMMENDED_ANILIST)
+        section_data['recommended']['trakt']['selectedLists'] = app.CACHE_RECOMMENDED_TRAKT_LISTS
+        section_data['recommended']['trakt']['availableLists'] = TraktPopular.CATEGORIES
 
         section_data['versionNotify'] = bool(app.VERSION_NOTIFY)
         section_data['autoUpdate'] = bool(app.AUTO_UPDATE)