+++ tests/apiv2/test_config.py
@@ -15,6 +15,7 @@ from medusa.indexers.config import INDEXER_TVDBV2
 from medusa.network_timezones import app_timezone
 from medusa.sbdatetime import date_presets, time_presets
 from medusa.schedulers.utils import all_schedulers
+from medusa.show.recommendations.trakt import TraktPopular
 
 from tests.apiv2.conftest import TEST_API_KEY
 
@@ -87,6 +88,10 @@ def config_main(monkeypatch, app_config):
 
     section_data['indexerDefaultLanguage'] = app.INDEXER_DEFAULT_LANGUAGE
     section_data['showUpdateHour'] = int_default(app.SHOWUPDATE_HOUR, app.DEFAULT_SHOWUPDATE_HOUR)
+    section_data['recommendedShowUpdateHour'] = int_default(
+        app.RECOMMENDED_SHOW_UPDATE_HOUR, app.DEFAULT_RECOMMENDED_SHOW_UPDATE_HOUR
+    )
+
     section_data['indexerTimeout'] = int_default(app.INDEXER_TIMEOUT, 20)
     section_data['indexerDefault'] = app.INDEXER_DEFAULT
 
@@ -94,6 +99,15 @@ def config_main(monkeypatch, app_config):
     section_data['plexFallBack']['enable'] = bool(app.FALLBACK_PLEX_ENABLE)
     section_data['plexFallBack']['notifications'] = bool(app.FALLBACK_PLEX_NOTIFICATIONS)
     section_data['plexFallBack']['timeout'] = int(app.FALLBACK_PLEX_TIMEOUT)
+
+    section_data['recommended'] = {'cache': {}, 'trakt': {}}
+    section_data['recommended']['cache']['shows'] = bool(app.CACHE_RECOMMENDED_SHOWS)
+    section_data['recommended']['cache']['trakt'] = bool(app.CACHE_RECOMMENDED_TRAKT)
+    section_data['recommended']['cache']['imdb'] = bool(app.CACHE_RECOMMENDED_IMDB)
+    section_data['recommended']['cache']['anidb'] = bool(app.CACHE_RECOMMENDED_ANIDB)
+    section_data['recommended']['cache']['anilist'] = bool(app.CACHE_RECOMMENDED_ANILIST)
+    section_data['recommended']['trakt']['selectedLists'] = app.CACHE_RECOMMENDED_TRAKT_LISTS
+    section_data['recommended']['trakt']['availableLists'] = TraktPopular.CATEGORIES
 
     section_data['versionNotify'] = bool(app.VERSION_NOTIFY)
     section_data['autoUpdate'] = bool(app.AUTO_UPDATE)