+++ themes-default/slim/src/components/config-notifications.vue
@@ -1081,7 +1081,6 @@ export default {
             });
         }
     },
-
     beforeMount() {
         // Wait for the next tick, so the component is rendered
         this.$nextTick(() => {